Fix bug in delete method
The original implementation of the delete method decremented the pointer to the
'last' pointer element each time it was called. This was not an issue since the
other methods only refer don't refer to elements in the backing array.

It does become an issue when the number of elements are less than three as the
getLargestChildIndex method is comparing the element in index 1 to a value that
is no longer part of the priority heap in index 2. If the value in index 2 is
larger than the value in index 1, it will be returned instead.

Zeroing out array indexs as they are deleted solves this issue.